requirements.txt
python-telegram-bot
requests
PyYAML
Pillow